Evenly space slash separator
Make the slash separating page heading levels evenly spaced on the
left and right (currently there is a space after it, but not before).
Do this via some css fudging, so we don't need to change ikiwiki
templates or anything.

The degree to which the uneven slashes bother me is probably
irrational, so if anyone strongly disagrees with this, feel free to
revert and I can just do this locally.